requirements.txt
pytest
tox
coverage
pytest-cov
pytest-flask
Flask
flask_restful
flask_restplus
swagger-ui-py
Flask-Caching
redis
flask-jwt-extended
passlib
flask-sqlalchemy
psycopg2-binary
Flask-Migrate
flask-migrate
Flask-Restless==0.16.0
Flask-JWT==0.3.2
Flask-Security==1.7.5
Flask-Admin==1.5.1
passlib
jwt
greenlet==0.4.15
gevent==1.5.0
gunicorn
Werkzeug==0.16.1
flask_bcrypt
flask_cors
dependency_injector
marshmallow
flake8
mock
pytest_mock
healthcheck